chore: cycle list layout consistency
